9043/refactor: cleaned up ProcessPaymentServiceImpl extracted some methods
