+++ service-api/src/main/java/greencity/service/ubs/payment/ProcessPaymentService.java
@@ -3,10 +3,10 @@ package greencity.service.ubs.payment;
 import greencity.dto.order.OrderResponseDto;
 import greencity.dto.order.OrderWayForPayClientDto;
 import greencity.dto.order.PaymentSystemResponse;
+import greencity.dto.payment.PaymentWayForPayRequestDto;
 import greencity.entity.order.Order;
 import greencity.entity.user.User;
 
-//TODO add docs
 //TODO add test
 public interface ProcessPaymentService {
     /**
@@ -40,5 +40,16 @@ public interface ProcessPaymentService {
      */
     PaymentSystemResponse processOrder(String userUuid, OrderWayForPayClientDto dto);
 
+    /**
+     * Forms a payment link for redirecting the user to WayForPay checkout.
+     * - Increments the order’s counter for payment attempts. <br>
+     * - Creates a {@link PaymentWayForPayRequestDto} request with encoded order reference. <br>
+     * - Sends the request to WayForPay checkout client. <br>
+     * - Extracts and returns the redirect URL from the WayForPay response.
+     *
+     * @param order           the {@link Order} for which the payment link is generated
+     * @param sumToPayInCoins the amount to be paid in coins
+     * @return the checkout URL where the client should be redirected to complete payment
+     */
     String formedLink(Order order, long sumToPayInCoins);
 }
