+++ service/src/main/java/greencity/service/ubs/order/OrderServiceImpl.java
@@ -1,0 +1,174 @@
+package greencity.service.ubs.order;
+
+import static greencity.constant.ErrorMessage.TARIFF_FOR_BAGS_AT_LOCATION_NOT_EXIST;
+import static greencity.constant.ErrorMessage.TOO_MUCH_POINTS_FOR_ORDER;
+import static java.util.Objects.nonNull;
+import greencity.constant.AppConstant;
+import greencity.dto.bag.BagDto;
+import greencity.dto.order.OrderResponseDto;
+import greencity.entity.order.Certificate;
+import greencity.entity.order.ChangeOfPoints;
+import greencity.entity.order.Order;
+import greencity.entity.order.OrderBag;
+import greencity.entity.order.Payment;
+import greencity.entity.order.TariffsInfo;
+import greencity.entity.user.User;
+import greencity.entity.user.ubs.UBSuser;
+import greencity.enums.BonusReason;
+import greencity.enums.OrderPaymentStatus;
+import greencity.enums.OrderStatus;
+import greencity.enums.PaymentStatus;
+import greencity.exceptions.BadRequestException;
+import greencity.exceptions.NotFoundException;
+import greencity.repository.OrderRepository;
+import greencity.repository.TariffsInfoRepository;
+import greencity.service.ubs.calculator.BagCalculatorService;
+import greencity.service.ubs.calculator.CertificateCalculatorService;
+import greencity.service.ubs.calculator.PaymentCalculatorService;
+import greencity.service.ubs.calculator.PointCalculatorService;
+import greencity.util.PointsUtils;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections4.CollectionUtils;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class OrderServiceImpl implements OrderService {
+    private final OrderRepository orderRepository;
+    private final TariffsInfoRepository tariffsInfoRepository;
+    private final PointCalculatorService pointCalculatorService;
+    private final BagCalculatorService bagCalculatorService;
+    private final CertificateCalculatorService certificateCalculatorService;
+    private final PaymentCalculatorService paymentCalculatorService;
+    private final PointsUtils pointsUtils;
+
+    @Override
+    public Order formAndSaveOrderRequest(OrderResponseDto dto, Order order, User currentUser, UBSuser userData) {
+        TariffsInfo tariffsInfo = findTariffsInfoByBagIdsWithinLocation(
+            getBagIds(dto.getBags()), dto.getLocationId());
+        List<OrderBag> bagsOrdered = new ArrayList<>();
+        long sumToPayInCoinsWithoutDiscount = bagCalculatorService
+            .prepareBagsAndCalculateTotal(bagsOrdered, dto.getBags(), tariffsInfo);
+
+        pointsUtils.checkIfUserHaveEnoughPoints(currentUser.getCurrentPoints(), dto.getPointsToUse());
+        long sumToPayInCoins = pointCalculatorService
+            .reduceOrderSumDueToUsedPoints(sumToPayInCoinsWithoutDiscount, dto.getPointsToUse());
+        if (sumToPayInCoinsWithoutDiscount == sumToPayInCoins) {
+            order.setPointsToUse(0);
+            dto.setPointsToUse(0);
+        }
+
+        Set<Certificate> orderCertificates = new HashSet<>();
+        sumToPayInCoins =
+            certificateCalculatorService.applyCertificatesToOrder(
+                dto, orderCertificates, order, sumToPayInCoins);
+        if (sumToPayInCoins <= 0) {
+            dto.setShouldBePaid(false);
+        }
+        return formAndSaveOrder(order, orderCertificates, bagsOrdered, userData, currentUser, sumToPayInCoins,
+            tariffsInfo);
+    }
+
+    @Override
+    public void transferUserPointsToOrder(Order order, Integer pointsToUse) {
+        if (pointsToUse <= 0) {
+            return;
+        }
+
+        User user = order.getUser();
+        pointsUtils.checkIfUserHaveEnoughPoints(user.getCurrentPoints(), pointsToUse);
+
+        int maxPointsToTransfer = countAmountToPayForOrder(order);
+        if (pointsToUse > maxPointsToTransfer) {
+            throw new BadRequestException(TOO_MUCH_POINTS_FOR_ORDER + maxPointsToTransfer);
+        }
+
+        order.setPointsToUse(order.getPointsToUse() + pointsToUse);
+        user.setCurrentPoints(user.getCurrentPoints() - pointsToUse);
+        user.getChangeOfPointsList()
+            .add(ChangeOfPoints.builder()
+                .user(user)
+                .amount(-pointsToUse)
+                .date(LocalDateTime.now())
+                .order(order)
+                .reason(BonusReason.DEBIT_PAYMENT)
+                .build());
+
+        orderRepository.save(order);
+    }
+
+    private TariffsInfo findTariffsInfoByBagIdsWithinLocation(List<Integer> bagIds, Long locationId) {
+        return tariffsInfoRepository.findTariffsInfoByBagIdAndLocationId(bagIds, locationId)
+            .orElseThrow(
+                () -> new NotFoundException(String.format(TARIFF_FOR_BAGS_AT_LOCATION_NOT_EXIST, bagIds, locationId)));
+    }
+
+    private List<Integer> getBagIds(List<BagDto> dto) {
+        return dto.stream()
+            .map(BagDto::getId)
+            .toList();
+    }
+
+    private Order formAndSaveOrder(
+        Order order, Set<Certificate> orderCertificates, List<OrderBag> bagsOrdered,
+        UBSuser userData, User currentUser, long sumToPayInCoins, TariffsInfo tariffsInfo
+    ) {
+        order.setTariffsInfo(tariffsInfo);
+        order.setCertificates(orderCertificates);
+        order.setOrderBags(bagsOrdered);
+        order.setUbsUser(userData);
+        order.setUser(currentUser);
+        order.setSumTotalAmountWithoutDiscounts(
+            paymentCalculatorService.calculateOrderSumWithoutDiscounts(bagsOrdered));
+        order.setCounterOrderPaymentId(order.getCounterOrderPaymentId() + 1);
+        setOrderPaymentStatus(order, sumToPayInCoins);
+
+        Payment payment = Payment.builder()
+            .amount(sumToPayInCoins)
+            .orderStatus(OrderStatus.FORMED)
+            .currency("UAH")
+            .paymentStatus(PaymentStatus.UNPAID)
+            .settlementDate(LocalDate.now().format(DateTimeFormatter.ISO_LOCAL_DATE))
+            .order(order).build();
+
+        if (order.getPayment() == null) {
+            order.setPayment(new ArrayList<>());
+        }
+        order.getPayment().add(payment);
+        return orderRepository.save(order);
+    }
+
+    private void setOrderPaymentStatus(Order order, long sumToPay) {
+        if (sumToPay <= 0) {
+            order.setOrderPaymentStatus(OrderPaymentStatus.PAID);
+        } else {
+            order.setOrderPaymentStatus(
+                order.getPointsToUse() > 0 || CollectionUtils.isNotEmpty(order.getCertificates())
+                    ? OrderPaymentStatus.HALF_PAID
+                    : OrderPaymentStatus.UNPAID);
+        }
+    }
+
+    private int countAmountToPayForOrder(Order order) {
+        int certificatesAmount = nonNull(order.getCertificates())
+            ? order.getCertificates().stream()
+            .map(Certificate::getPoints)
+            .reduce(0, Integer::sum)
+            : 0;
+        return -order.getPointsToUse() - certificatesAmount
+            + BigDecimal.valueOf(order.getSumTotalAmountWithoutDiscounts())
+            .movePointLeft(AppConstant.TWO_DECIMALS_AFTER_POINT_IN_CURRENCY)
+            .setScale(0, RoundingMode.UP).intValue();
+    }
+}
